src/requirements.txt
numpy==1.13.1
scipy==0.19.1
tensorflow-gpu==1.2.0
